Separating layers.api.js file into multiple files Adding support for @time=... (current timestamp of a layer) in the URL
