+++ src/api/layers/LayerTypes.enum.js
@@ -1,0 +1,11 @@
+/**
+ * @readonly
+ * @enum {String}
+ */
+const LayerTypes = {
+    WMTS: 'wmts',
+    WMS: 'wms',
+    GEOJSON: 'geojson',
+    AGGREGATE: 'aggregate',
+}
+export default LayerTypes
