+++ src/modules/store/plugins/load-layersconfig-on-lang-change.js
@@ -1,5 +1,5 @@
 import { SET_LANG_MUTATION_KEY } from '@/modules/i18n/store/i18n.store'
-import loadLayersConfigFromBackend from '@/api/layers.api'
+import { loadLayersConfigFromBackend } from '@/api/layers/layers.api'
 import loadTopicsFromBackend, { loadTopicTreeForTopic } from '@/api/topics.api'
 import log from '@/utils/logging'
