+++ src/router/storeSync/abstractParamConfig.class.js
@@ -102,4 +102,29 @@ export default class AbstractParamConfig {
             }
         }
     }
+
+    /**
+     * Sets the store values according to the URL. Returns a promise that will resolve when the
+     * store is up to date.
+     *
+     * @param {Vuex.Store} store
+     * @param {Object | String | Number | Boolean | null} query The value found in the query
+     * @returns {Promise<any>}
+     */
+    populateStoreWithQueryValue(store, query) {
+        return new Promise((resolve, reject) => {
+            if (query && store && this.setValuesInStore) {
+                const promiseSetValuesInStore = this.setValuesInStore(store, query)
+                if (promiseSetValuesInStore) {
+                    promiseSetValuesInStore.then(() => {
+                        resolve()
+                    })
+                } else {
+                    resolve()
+                }
+            } else {
+                reject('Query, store or setter functions is not set')
+            }
+        })
+    }
 }
